+++ source/extensions/omni.isaac.orbit/omni/isaac/orbit/sensors/camera/utils.py
@@ -176,7 +176,7 @@ def create_pointcloud_from_rgbd(
 
     The ``rgb`` attribute is used to resolve the corresponding point's color:
 
-    - If a ``np.array``/``wp.arrray``/``torch.tensor`` of shape (H, W, 3), then the corresponding channels encode RGB values.
+    - If a ``np.array``/``wp.array``/``torch.tensor`` of shape (H, W, 3), then the corresponding channels encode RGB values.
     - If a tuple, then the point cloud has a single color specified by the values (r, g, b).
     - If :obj:`None`, then default color is white, i.e. (0, 0, 0).